requirements.txt
torch==1.10.2
pandas==1.4.0
rdkit-pypi==2021.9.4
scikit-learn==1.0.2
pytorch-lightning==1.5.10